Refactor monitoring UI to use React hooks
